+++ src/com/android/mms/ui/MessageUtils.java
@@ -850,9 +850,9 @@ public class MessageUtils {
         return getLocalNumber(SubscriptionManager.getDefaultDataSubId());
     }
 
-    public static String getLocalNumber(long subId) {
+    public static String getLocalNumber(int subId) {
         sLocalNumber = MmsApp.getApplication().getTelephonyManager()
-            .getLine1NumberForSubscriber((int)subId);
+            .getLine1NumberForSubscriber(subId);
         return sLocalNumber;
     }
