+++ src/com/android/mms/ui/NoConfirmationSendService.java
@@ -65,7 +65,7 @@ public class NoConfirmationSendService extends IntentService {
             ComposeMessageActivity.log("Called to send SMS but no extras");
             return;
         }
-        long sub = extras.getLong(PhoneConstants.SUBSCRIPTION_KEY, SubscriptionManager
+        int sub = extras.getInt(PhoneConstants.SUBSCRIPTION_KEY, SubscriptionManager
                 .getDefaultSubId());
 
         String message = extras.getString(Intent.EXTRA_TEXT);
@@ -92,7 +92,7 @@ public class NoConfirmationSendService extends IntentService {
             // provider looks up the threadId based on the recipient(s).
             long threadId = 0;
             SmsMessageSender smsMessageSender = new SmsMessageSender(this, dests, message, threadId,
-                    SubscriptionManager.getPhoneId((int)sub));
+                    SubscriptionManager.getPhoneId(sub));
             try {
                 // This call simply puts the message on a queue and sends a broadcast to start
                 // a service to send the message. In queing up the message, however, it does
